fix : 커뮤니티 필터링, 알람 리스트 subview 등으로 분리
